# Backend/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title:  API vías clínicas
  description: Documentación de la API de la herramienta software de la gestión de vías clínicas
...
paths:
  /user:
    get:
      summary: Obtiene la lista de médicos
      tags: [Médicos]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de médicos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - nombre
                    - usuario
                    - contrasegna
                    - esAdmin
                  properties:
                    id:
                      type: integer
                      description: Identificador único del médico
                      example: 1
                    nombre:
                      type: string
                      description: Nombre del médico
                      example: Juan Perez
                    usuario:
                      type: string
                      description: Nombre de médico en el sistema
                      example: jperez
                    contrasegna:
                      type: string
                      description: Contraseña del médico en el sistema
                      example: password123
                    esAdmin:
                      type: boolean
                      description: Indica si el médico es administrador
                      example: false
        403:
          description: Forbidden
...
  /login:
    post:
      summary: Inicio sesión 
      tags: [Médicos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Identificador del paciente
                  example: jperez
                password: 
                  type: string
                  description: Contraseña del usuario
                  example: password123
      responses:
        200:
          description: Inicio de sesión correcto
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Bear token
        500:
          description: Error en el servidor
        403:
          description: Forbidden
        409:
          description: "Duplicate entry, unique constraint violated."
  /logout:
    post:
      summary: Cerrar sesión 
      tags: [Médicos]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Se ha cerrado la sesión correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Bear token
        500:
          description: Error en el servidor
...
  /patient:
    get:
      summary: Obtiene la lista de pacientes
      tags: [Pacientes]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de pacientes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - numhist
                    - nombre
                    - sexo
                  properties:
                    numhist:
                      type: integer
                      description: Número de historia clínica del paciente
                      example: 111111
                    nombre:
                      type: string
                      description: Nombre del paciente
                      example: Juan Perez
                    sexo:
                      type: boolean
                      description: Sexo del paciente, true si es hombre, false si es mujer
                      example: true
                    
                    
        403:
          description: Forbidden
...
  /operacion:
    get:
      summary: Obtiene la lista de todas las operaciones 
      tags: [Operaciones]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de operaciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - indice
                    - edad
                    - fecha
                    - etiqueta
                    - idvia
                    - idoperacion
                    - numhist
                  properties:
                    id:
                      type: integer
                      description: Identificador único de la operación
                      example: 1
                    indice:
                      type: integer
                      description: Índice de la operación
                      example: 21
                    edad:
                      type: integer
                      description: Edad del paciente en el momento de la operación
                      example: 21
                    fecha:
                      type: string
                      format: date-time
                      description: Fecha de la operación
                      example: 2024-11-07T00:00:00.000Z
                    etiqueta:
                      type: string
                      description: Descripción de la operación
                      example: artroplasita primera izquierda
                    idvia:
                      type: integer
                      description: Identificador de la vía clínica
                      example: 1
                    idoperacion:
                      type: integer
                      description: Identificador único de la operación
                      example: 1
                    numhist:
                      type: integer
                      description: Número de historia clínica del paciente
                      example: 111111
                    
                    
        403:
          description: Forbidden
...
  /vias:
      get:
        summary: Obtiene la lista de vías
        tags: [Vías clínicas]
        security:
          - bearerAuth: []
        responses:
          200:
            description: Lista de vías obtenida exitosamente
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    required:
                      - id
                      - nombre
                    properties:
                      id:
                        type: integer
                        description: Identificador único de la vía clíncia
                        example: 1
                      nombre:
                        type: string
                        description: Nombre de la vía clínica
                        example: via rodilla
          403:
            description: Forbidden
...
  /tareas:
      get:
        summary: Obtiene las tareas
        tags: [Tareas]
        security:
          - bearerAuth: []
        responses:
          200:
            description: Lista de tareas obtenida exitosamente
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    required:
                      - id
                      - descripción
                      - id_fase
                    properties:
                      id:
                        type: integer
                        description: Identificador único de la tarea
                        example: 1
                      descripcion:
                        type: string
                        description: Descripción de la tarea
                        example: Realización de radiografías y telemétricas
                      id_fase:
                        type: integer
                        description: Identificador de la fase a la que pertenece la tarea
                        example: 1
                      tipo:
                        type: string
                        description: Tipo de tarea
                        example: booleano
          403:
            description: Forbidden
      post:
        summary: Crea una nueva tarea
        tags: 
          - Tareas
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - descripcion
                  - id_fase
                  - tipo
                properties:
                  descripcion:
                    type: string
                    description: Descripción de la tarea
                    example: Realización de radiografías y telemétricas
                  id_fase:
                    type: integer
                    description: Identificador de la fase a la que pertenece la tarea
                    example: 1
                  tipo:
                    type: string
                    description: Tipo de tarea
                    example: booleano || opciones
                  opciones:
                    type: array
                    items:
                      type: string
                    description: Opciones de la tarea
                    example: [si, no, tal vez]
                  
        responses:
          201:
            description: Tarea creada exitosamente
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      description: Mensaje de éxito
                      example: Tarea creada correctamente
                    body:
                      type: object
                      properties:
                        tarea:
                          type: object
                          properties:
                            id:
                              type: integer
                              description: Identificador único de la tarea
                              example: 1
                            descripcion:
                              type: string
                              description: Descripción de la tarea
                              example: Realización de radiografías y telemétricas
                            id_fase:
                              type: integer
                              description: Identificador de la fase a la que pertenece la tarea
                              example: 1
                            tipo:
                              type: string
                              description: Tipo de tarea
                              example: booleano 

          500:
            description: Error en el servidor
...
  /fases:
      get:
        summary: Obtiene las fases
        tags: [Fases]
        security:
          - bearerAuth: []
        responses:
          200:
            description: Lista de fases obtenida exitosamente
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    required:
                      - id
                      - nombre
                      - idvia
                    properties:
                      id:
                        type: integer
                        description: Identificador único de la fase
                        example: 1
                      nombre:
                        type: string
                        description: Nombre de la fase
                        example: Preingreso
                      idvia:
                        type: integer
                        description: Identificador de la via a la que pertenece la fase
                        example: 1
          403:
            description: Forbidden
      post:
        summary: Crea una nueva fase
        tags: 
          - Fases
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - nombre
                  - idvia
                properties:
                  nombre:
                    type: string
                    description: Nombre de la fase
                    example: Preingreso
                  idvia:
                    type: integer
                    description: Identificador de la via a la que pertenece la fase
                    example: 1                  
        responses:
          201:
            description: Fase creada exitosamente
            content:
              application/json:
                schema:
                  type: object
                  properties:
                          id:
                            type: integer
                            description: Identificador único de la fase
                            example: 1
                          nombre:
                            type: string
                            description: Nombre de la fase
                            example: Preingreso
                          idvia:
                            type: integer
                            description: Identificador de la via a la que pertenece la fase
                            example: 1
          500:
            description: Error en el servidor